4289 client - Fix toast throttling and extract shared utilities
