+++ client/src/config/useFetchInterceptor.ts
@@ -1,41 +1,12 @@
 import {useWorkspaceStore} from '@/pages/automation/stores/useWorkspaceStore';
 import {useAuthenticationStore} from '@/shared/stores/useAuthenticationStore';
 import {getCookie} from '@/shared/util/cookie-utils';
+import {shouldShowToast, showErrorToast} from '@/shared/util/toast-throttle';
 import fetchIntercept from 'fetch-intercept';
 import {useEffect, useRef} from 'react';
 import {useNavigate} from 'react-router-dom';
-import {toast} from 'sonner';
 
-const TOAST_COOLDOWN_MS = 10_000;
-const recentToastTimestamps = new Map<string, number>();
-
-export function clearRecentToasts() {
-    recentToastTimestamps.clear();
-}
-
-function shouldShowToast(toastId: string): boolean {
-    const now = Date.now();
-
-    for (const [id, timestamp] of recentToastTimestamps) {
-        if (now - timestamp >= TOAST_COOLDOWN_MS) {
-            recentToastTimestamps.delete(id);
-        }
-    }
-
-    const lastShown = recentToastTimestamps.get(toastId);
-
-    if (lastShown !== undefined && now - lastShown < TOAST_COOLDOWN_MS) {
-        return false;
-    }
-
-    recentToastTimestamps.set(toastId, now);
-
-    return true;
-}
-
-function showErrorToast(toastId: string, title: string, options?: {description?: string}) {
-    toast.error(title, {...options, id: toastId});
-}
+export {clearRecentToasts} from '@/shared/util/toast-throttle';
 
 export default function useFetchInterceptor() {
     const clearAuthentication = useAuthenticationStore((state) => state.clearAuthentication);
@@ -93,10 +64,6 @@ export default function useFetchInterceptor() {
 
                 const toastId = `${new URL(response.url).pathname}-${response.status}`;
 
-                if (!shouldShowToast(toastId)) {
-                    return response;
-                }
-
                 if (response.url.includes('/graphql')) {
                     const clonedResponse = response.clone();
 
@@ -104,21 +71,33 @@ export default function useFetchInterceptor() {
                         .json()
                         .then((data: {errors?: Array<{message?: string}>}) => {
                             if (data.errors?.length) {
+                                if (!shouldShowToast(toastId)) {
+                                    return;
+                                }
+
                                 const errorMessage = [
                                     ...new Set(data.errors.map((error) => error.message || 'Unknown error')),
                                 ].join('\n');
 
                                 showErrorToast(toastId, 'Error', {description: errorMessage});
                             } else if (response.status < 200 || response.status > 299) {
-                                showErrorToast(toastId, `Request failed with status ${response.status}`);
+                                if (shouldShowToast(toastId)) {
+                                    showErrorToast(toastId, `Request failed with status ${response.status}`);
+                                }
                             }
                         })
                         .catch(() => {
                             if (response.status < 200 || response.status > 299) {
-                                showErrorToast(toastId, `Request failed with status ${response.status}`);
+                                if (shouldShowToast(toastId)) {
+                                    showErrorToast(toastId, `Request failed with status ${response.status}`);
+                                }
                             }
                         });
                 } else if (response.status < 200 || response.status > 299) {
+                    if (!shouldShowToast(toastId)) {
+                        return response;
+                    }
+
                     const clonedResponse = response.clone();
 
                     clonedResponse
