+++ client/src/ee/pages/embedded/automation-workflows/workflow-builder/config/useFetchInterceptor.ts
@@ -1,40 +1,11 @@
 import {useWorkspaceStore} from '@/pages/automation/stores/useWorkspaceStore';
 import {useAuthenticationStore} from '@/shared/stores/useAuthenticationStore';
 import {useEnvironmentStore} from '@/shared/stores/useEnvironmentStore';
+import {shouldShowToast, showErrorToast} from '@/shared/util/toast-throttle';
 import fetchIntercept from 'fetch-intercept';
 import {useEffect, useRef} from 'react';
-import {toast} from 'sonner';
 
-const TOAST_COOLDOWN_MS = 10_000;
-const recentToastTimestamps = new Map<string, number>();
-
-export function clearRecentToasts() {
-    recentToastTimestamps.clear();
-}
-
-function shouldShowToast(toastId: string): boolean {
-    const now = Date.now();
-
-    for (const [id, timestamp] of recentToastTimestamps) {
-        if (now - timestamp >= TOAST_COOLDOWN_MS) {
-            recentToastTimestamps.delete(id);
-        }
-    }
-
-    const lastShown = recentToastTimestamps.get(toastId);
-
-    if (lastShown !== undefined && now - lastShown < TOAST_COOLDOWN_MS) {
-        return false;
-    }
-
-    recentToastTimestamps.set(toastId, now);
-
-    return true;
-}
-
-function showErrorToast(toastId: string, title: string, options?: {description?: string}) {
-    toast.error(title, {...options, id: toastId});
-}
+export {clearRecentToasts} from '@/shared/util/toast-throttle';
 
 export default function useFetchInterceptor() {
     const clearAuthentication = useAuthenticationStore((state) => state.clearAuthentication);
@@ -92,34 +63,13 @@ export default function useFetchInterceptor() {
                     return response;
                 }
 
-                const toastId = `${new URL(response.url).pathname}-${response.status}`;
+                if (response.status < 200 || response.status > 299) {
+                    const toastId = `${new URL(response.url).pathname}-${response.status}`;
 
-                if (!shouldShowToast(toastId)) {
-                    return response;
-                }
+                    if (!shouldShowToast(toastId)) {
+                        return response;
+                    }
 
-                if (response.url.includes('/graphql')) {
-                    const clonedResponse = response.clone();
-
-                    clonedResponse
-                        .json()
-                        .then((data: {errors?: Array<{message?: string}>}) => {
-                            if (data.errors?.length) {
-                                const errorMessage = [
-                                    ...new Set(data.errors.map((error) => error.message || 'Unknown error')),
-                                ].join('\n');
-
-                                showErrorToast(toastId, 'Error', {description: errorMessage});
-                            } else if (response.status < 200 || response.status > 299) {
-                                showErrorToast(toastId, `Request failed with status ${response.status}`);
-                            }
-                        })
-                        .catch(() => {
-                            if (response.status < 200 || response.status > 299) {
-                                showErrorToast(toastId, `Request failed with status ${response.status}`);
-                            }
-                        });
-                } else if (response.status < 200 || response.status > 299) {
                     const clonedResponse = response.clone();
 
                     clonedResponse
